% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CycloneEnergy.R
\name{ace_of_storm}
\alias{ace_of_storm}
\title{Compute the accumulated cyclone energy of the storm}
\usage{
ace_of_storm(storm_data, storm_id)
}
\arguments{
\item{storm_data}{data frame of storm data}

\item{id}{id of the storm to calculate}
}
\value{
Energy accumulated cyclone energy of this storm
}
\description{
A function that returns the accumulated cyclone energy of a given storm
}
\examples{
ace= ace_of_storm(data,"AL122007")
}
