% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dataset.R
\name{format}
\alias{format}
\title{Format the hurdat dataset in the data directory}
\usage{
format(dat)
}
\arguments{
\item{dat}{the text file to be formatted}
}
\value{
the formatted data frame save as RData
}
\description{
Get a data frame with the following qualities:
no storm id rows.
instead of storm id rows, an id column to signify which storm id each row corresponds to.
a name column for the name of the storm
missing value codes replaced with NAs
meaningful column names. (see documentation)
numeric latitude and longitude columns, in degrees north and degrees east, respectively.
}
\examples{
dat <- read.csv("hurdat2-1851-2022-040723.txt", header= FALSE)
format(dat)
}
