% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StormTrack30m.R
\name{interpolate_storm_track_30m}
\alias{interpolate_storm_track_30m}
\title{Interpolate storm tracks with 30min interval}
\usage{
interpolate_storm_track_30m(storm_data)
}
\arguments{
\item{storm_data}{the data frame to be modified}
}
\value{
the interpolated data frame
}
\description{
A function for interpolating a storm track to 30 minute increments.
Only longitude and latitude will be filled with interpolated value.
Other columns will be filled with the same value in the preceding row
}
\examples{
new_df <- interpolate_storm_track_30m(data[1:200,])
head(new_df)
}
