% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapAStorm.R
\name{plot_storm_size}
\alias{plot_storm_size}
\title{Plot the position and size of a storm}
\usage{
plot_storm_size(storm_data, row_index)
}
\arguments{
\item{storm_data}{data frame of storm data}

\item{row_index}{row index of the storm to plot}
}
\description{
A function to plot a map that contains a polygon of the storm from the given row.
Use the 34(blue), 50(green), and 64(orange) knot extent variables to determine the polygon.
1 nautical mile = 1/60 degrees of latitude
1 nautical mile = 1/(60 * cos(latitude)) degrees of longitude are used for approximations.
}
\examples{
plot_storm_size(your_storm_data, 1)
}
