% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StormsMap.R
\name{plot_storm_tracks}
\alias{plot_storm_tracks}
\title{Plot storm tracks of selsected storms}
\usage{
plot_storm_tracks(storm_data, storm_ids)
}
\arguments{
\item{ids}{id of the storm}
}
\value{
length of subset
}
\description{
A function for plotting a map of storm tracks for a selection of storms ids in a vector.
The map includes country and US state boundaries.
}
\examples{
plot_storm_tracks(your_storm_data, c("storm_id_1", "storm_id_2"))
}
