% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntoUSA.R
\name{storm_US_landfall}
\alias{storm_US_landfall}
\title{If the storm landed on USA}
\usage{
storm_US_landfall(storm_data, storm_id)
}
\arguments{
\item{storm_data}{data frame of storm data}

\item{storm_id}{id of this storm}
}
\value{
true or false
}
\description{
A function for determining whether the storm made landfall in the continental United States
Note:continental US includes Alaska(?), but excludes Hawaii and the U.S. territories such as Puerto Rico and the U.S. Virgin Islands.
No use of the unreliable record identifier variable for this part.
Return true if it landed in the continental United States, false otherwise.
}
\examples{
IfLandfall<-storm_US_landfall(data,"AL122007")
}
